# wasm: experiment with one off eval function