--- a/capabilities.json
+++ b/capabilities.json
@@ -3503,6 +3503,10 @@
     {
       "version": 1,
       "minor_version": 1
+    },
+    {
+      "version": 1,
+      "minor_version": 2
     }
   ]
 }