--- a/internal/compiler/wasm/opa/callgraph.csv
+++ b/internal/compiler/wasm/opa/callgraph.csv
@@ -200,6 +200,12 @@ opa_cmp_lt,opa_boolean
 opa_cmp_lte,opa_value_compare
 opa_cmp_lte,opa_boolean
 opa_eval_ctx_new,opa_malloc
+opa_eval,opa_abort
+opa_eval,opa_heap_ptr_set
+opa_eval,opa_value_parse
+opa_eval,eval
+opa_eval,opa_value_dump
+opa_eval,opa_json_dump
 __force_import_opa_builtins,opa_builtin0
 __force_import_opa_builtins,opa_builtin1
 __force_import_opa_builtins,opa_builtin2