--- a/rego/testdata/t.rego
+++ b/rego/testdata/t.rego
@@ -1,0 +1,7 @@
+package test
+
+default allow = false
+
+allow {
+    input.x == "x"
+}