This increases performance quite a bit:

```
name         old time/op    new time/op    delta
WasmRego-16    76.8µs ± 3%    20.6µs ± 1%  -73.23%  (p=0.008 n=5+5)

name         old alloc/op   new alloc/op   delta
WasmRego-16    6.05kB ± 0%    2.30kB ± 0%  -61.99%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
WasmRego-16       267 ± 0%        75 ± 0%  -71.87%  (p=0.008 n=5+5)
```
for the simple [benchmark here](https://github.com/open-policy-agent/opa/blob/27c8d75d3528dc8a04e716d5455f97ee7fb537bc/internal/wasm/sdk/opa/opa_bench_test.go#L20)